% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeSeatrackPassword.R
\name{changeSeatrackPassword}
\alias{changeSeatrackPassword}
\title{changeSeatrackPassword}
\usage{
changeSeatrackPassword(password)
}
\arguments{
\item{password}{Your new password (character string).}
}
\value{
Null
}
\description{
Changes the password for a user in the Seatrack database. Since the passwords for the file archive are fetched from the database,
this also affects the file archive.
}
\examples{
\dontrun{
changeSeatrackPassword("newPassword")
}

}
