% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackRdb-package.R
\docType{package}
\name{seatrackRdb-package}
\alias{seatrackRdb}
\alias{seatrackRdb-package}
\title{seatrackRdb: Interface functions for the Seatrack database}
\description{
R utility functions for interaction with the seatrack database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ninanor.github.io/seatrack-db/}
}

}
\author{
\strong{Maintainer}: Jens Åström \email{jens.astrom@nina.no} (\href{https://orcid.org/0000-0002-6114-0440}{ORCID})

Authors:
\itemize{
  \item Julian Evans \email{julian.evans@nina.nl}
}

}
\keyword{internal}
