% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePostable.R
\name{writePostable}
\alias{writePostable}
\title{Update the positions.postable}
\usage{
writePostable(positionData)
}
\arguments{
\item{positionData}{A list of position data to be read into the postable in the database. Usually created by \code{loadPosdata}.}
}
\value{
Data frame.
}
\description{
This is a convenience function that writes to the "positions.postable" table, the main table for the position data.
}
\examples{
\dontrun{
connectSeatrack(Username = "testreader", Password = "testreader")

 files<-c("posdata_FULGLA_eynhallow_2014",
          "posdata_FULGLA_eynhallow_2013",
          "posdata_FULGLA_eynhallow_2012",
          "posdata_FULGLA_eynhallow_2011",
          "posdata_FULGLA_eynhallow_2010",
          "posdata_FULGLA_eynhallow_2009",
          "posdata_FULGLA_eynhallow_2007"
          )

toImport <- loadPosdata(files)

summary(toImport)

writePostable(toImport)

}
}
