% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectSeatrack.R
\name{getSeatrackConnection}
\alias{getSeatrackConnection}
\title{Return ongoing database connection}
\usage{
getSeatrackConnection()
}
\description{
Utility function to access the seatrack db connection
}
